CalculatePropertyTax
This Java program is a Property Tax Calculator with a simple graphical interface. Users can enter a property value, and the program will calculate and show the assessment value (60% of the property value) and the property tax (0.64% of the assessment value). It also handles errors if the user enters invalid data and has a button to clear the input and results for new calculations. The interface is easy to use and provides helpful tips for the user.